+++ src/backend/ipcHandlers/index.ts
@@ -1,1 +1,2 @@
 import './launchdarkly'
+import './checkDiskSpace'
