+++ src/backend/main.ts
@@ -3,7 +3,6 @@ import { initImagesCache } from './images_cache'
 import { downloadAntiCheatData } from './anticheat/utils'
 import {
   AppSettings,
-  DiskSpaceData,
   ExecResult,
   GamepadInputEvent,
   GameSettings,
@@ -28,14 +27,9 @@ import {
 } from 'electron'
 import 'backend/updater'
 import { autoUpdater } from 'electron-updater'
-import { cpus, platform, release } from 'os'
-import { platform as platformNode } from 'process'
-import { promisify } from 'util'
-import { execFile } from 'child_process'
+import { cpus, platform } from 'os'
 import {
-  access,
   appendFileSync,
-  constants,
   existsSync,
   readdirSync,
   readFileSync,
@@ -47,9 +41,6 @@ import {
 import * as LDElectron from 'launchdarkly-electron-client-sdk'
 import Backend from 'i18next-fs-backend'
 import i18next from 'i18next'
-import checkDiskSpace, {
-  getFirstExistingParentPath
-} from '@hyperplay/check-disk-space'
 import { DXVK, Winetricks } from './tools'
 import { GameConfig } from './game_config'
 import { GlobalConfig } from './config'
@@ -63,7 +54,6 @@ import {
   clearCache,
   downloadDefaultWine,
   execAsync,
-  getFileSize,
   getGOGdlBin,
   getGogdlVersion,
   getLegendaryBin,
@@ -712,61 +702,6 @@ ipcMain.on('unlock', () => {
       powerSaveBlocker.stop(powerId)
     }
   }
-})
-
-ipcMain.handle('checkDiskSpace', async (event, folder) => {
-  const deps = {
-    platform: platformNode,
-    release: release(),
-    fsAccess: promisify(access),
-    pathNormalize: path.normalize,
-    pathSep: path.sep,
-    cpExecFile: promisify(execFile)
-  }
-  const parent = await getFirstExistingParentPath(folder, deps)
-  if (!parent) {
-    return {
-      free: 0,
-      diskSize: 0,
-      message: 'Path does not exist',
-      validPath: false
-    }
-  }
-  return new Promise<DiskSpaceData>((res) => {
-    access(parent, constants.W_OK, async (writeError) => {
-      const { free, size: diskSize } = await checkDiskSpace(folder).catch(
-        (checkSpaceError) => {
-          logError(
-            [
-              'Failed to check disk space for',
-              `"${folder}":`,
-              checkSpaceError.stack ?? `${checkSpaceError}`
-            ],
-            LogPrefix.Backend
-          )
-          return { free: 0, size: 0 }
-        }
-      )
-      if (writeError) {
-        logWarning(
-          [
-            'Cannot write to',
-            `"${folder}":`,
-            writeError.stack ?? `${writeError}`
-          ],
-          LogPrefix.Backend
-        )
-      }
-
-      const ret = {
-        free,
-        diskSize,
-        message: `${getFileSize(free)} / ${getFileSize(diskSize)}`,
-        validPath: !writeError
-      }
-      res(ret)
-    })
-  })
 })
 
 ipcMain.on('quit', async () => handleExit())
